Move tests from ForgeMod to DataGeneratorTest, fix test resources sourceset
