+++ src/generated_test/resources/.cache/cache
@@ -1,2 +1,12 @@
+e267e3cd85282291d97148a330e7857ba75df85d assets\data_gen_test\models\block\acacia_fence_gate.json
+a56c3472de2dd6b175f0926b295b7f0b9c92a0b7 assets\data_gen_test\models\block\acacia_fence_gate_open.json
+5c8a634491d6465c9da6983878ea22b71c0e0ae6 assets\data_gen_test\models\block\acacia_fence_gate_wall.json
+2ebae83ff7451e4c142cd46ae74b523d307618c8 assets\data_gen_test\models\block\acacia_fence_gate_wall_open.json
+611969c44ad660c1f6c12c9e5c5a7d784b7f7a22 assets\data_gen_test\models\block\acacia_fence_post.json
+e02938f13d9603e2386e35de87f1c4a7e6872747 assets\data_gen_test\models\block\acacia_fence_side.json
+42ed1755d3a63b3a718365b7c38dbe1998244b82 assets\data_gen_test\models\item\test_block_model.json
+726e1884e53151492ca05bd752c67031603570f6 assets\data_gen_test\models\item\test_generated_model.json
+596be057d7d1f03927f302ef69f3db794db0a2bb assets\minecraft\blockstates\acacia_fence.json
+79e7acd69879256b7f2f638f7f68e7774022b2a9 assets\minecraft\blockstates\acacia_fence_gate.json
 4fbaf6f4a3ea05cc071076e27f44ac81f9cc50e3 data\data_gen_test\advancements\conditional.json
 ed4cbf1a3a2f5d8969f6346fdc9acdbe81d0c919 data\data_gen_test\recipes\conditional.json
