+++ src/generated_test/resources/assets/data_gen_test/models/block/acacia_fence_gate.json
@@ -1,0 +1,6 @@
+{
+  "parent": "minecraft:block/template_fence_gate",
+  "textures": {
+    "texture": "minecraft:block/acacia_planks"
+  }
+}
