+++ src/generated_test/resources/assets/data_gen_test/models/block/acacia_fence_gate_wall_open.json
@@ -1,0 +1,6 @@
+{
+  "parent": "minecraft:block/template_fence_gate_wall_open",
+  "textures": {
+    "texture": "minecraft:block/acacia_planks"
+  }
+}
