+++ src/generated_test/resources/assets/data_gen_test/models/block/acacia_fence_side.json
@@ -1,0 +1,6 @@
+{
+  "parent": "minecraft:block/fence_side",
+  "textures": {
+    "texture": "minecraft:block/acacia_planks"
+  }
+}
