+++ src/generated_test/resources/assets/data_gen_test/models/item/test_generated_model.json
