+++ src/generated_test/resources/assets/minecraft/blockstates/acacia_fence.json
@@ -1,0 +1,47 @@
+{
+  "multipart": [
+    {
+      "apply": {
+        "model": "data_gen_test:block/acacia_fence_post"
+      }
+    },
+    {
+      "when": {
+        "north": "true"
+      },
+      "apply": {
+        "model": "data_gen_test:block/acacia_fence_side"
+      }
+    },
+    {
+      "when": {
+        "east": "true"
+      },
+      "apply": {
+        "model": "data_gen_test:block/acacia_fence_side",
+        "y": 90,
+        "uvlock": true
+      }
+    },
+    {
+      "when": {
+        "south": "true"
+      },
+      "apply": {
+        "model": "data_gen_test:block/acacia_fence_side",
+        "y": 180,
+        "uvlock": true
+      }
+    },
+    {
+      "when": {
+        "west": "true"
+      },
+      "apply": {
+        "model": "data_gen_test:block/acacia_fence_side",
+        "y": 270,
+        "uvlock": true
+      }
+    }
+  ]
+}
