+++ src/main/java/net/minecraftforge/common/ForgeMod.java
@@ -38,48 +38,36 @@ import org.apache.commons.lang3.tuple.Pair;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import net.minecraft.block.Block;
-import net.minecraft.block.Blocks;
-import net.minecraft.block.FenceBlock;
-import net.minecraft.block.FenceGateBlock;
 import net.minecraft.data.DataGenerator;
 import net.minecraft.item.crafting.IRecipe;
 import net.minecraft.item.crafting.IRecipeSerializer;
 import net.minecraft.nbt.CompoundNBT;
-import net.minecraft.util.Direction;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.world.storage.SaveHandler;
 import net.minecraft.world.storage.WorldInfo;
-import net.minecraftforge.client.model.generators.*;
-import net.minecraftforge.client.model.generators.ModelFile.UncheckedModelFile;
-import net.minecraftforge.common.crafting.*;
-import net.minecraftforge.common.crafting.conditions.*;
+import net.minecraftforge.common.crafting.CompoundIngredient;
+import net.minecraftforge.common.crafting.ConditionalRecipe;
+import net.minecraftforge.common.crafting.CraftingHelper;
+import net.minecraftforge.common.crafting.IngredientNBT;
+import net.minecraftforge.common.crafting.VanillaIngredientSerializer;
+import net.minecraftforge.common.crafting.conditions.AndCondition;
+import net.minecraftforge.common.crafting.conditions.FalseCondition;
+import net.minecraftforge.common.crafting.conditions.ItemExistsCondition;
+import net.minecraftforge.common.crafting.conditions.ModLoadedCondition;
+import net.minecraftforge.common.crafting.conditions.NotCondition;
+import net.minecraftforge.common.crafting.conditions.OrCondition;
+import net.minecraftforge.common.crafting.conditions.TagEmptyCondition;
+import net.minecraftforge.common.crafting.conditions.TrueCondition;
 import net.minecraftforge.common.data.ForgeBlockTagsProvider;
 import net.minecraftforge.common.data.ForgeItemTagsProvider;
 import net.minecraftforge.common.data.ForgeRecipeProvider;
 import net.minecraftforge.common.model.animation.CapabilityAnimation;
 import net.minecraftforge.energy.CapabilityEnergy;
 import net.minecraftforge.event.RegistryEvent;
-import net.minecraftforge.eventbus.api.IEventBus;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fluids.capability.CapabilityFluidHandler;
-import net.minecraftforge.fml.*;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.config.ModConfig;
-import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
-import net.minecraftforge.fml.event.lifecycle.FMLModIdMappingEvent;
 import net.minecraftforge.fml.event.lifecycle.GatherDataEvent;
-import net.minecraftforge.fml.event.server.FMLServerStartingEvent;
-import net.minecraftforge.fml.event.server.FMLServerStoppingEvent;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import net.minecraftforge.fluids.capability.CapabilityFluidHandler;
 import net.minecraftforge.items.CapabilityItemHandler;
-import net.minecraftforge.server.command.ConfigCommand;
-import net.minecraftforge.server.command.ForgeCommand;
-import net.minecraftforge.versions.forge.ForgeVersion;
-import net.minecraftforge.versions.mcp.MCPVersion;
-import org.apache.commons.lang3.tuple.Pair;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import net.minecraftforge.fml.common.Mod;
 import org.apache.logging.log4j.Marker;
 import org.apache.logging.log4j.MarkerManager;
 
@@ -193,11 +181,6 @@ public class ForgeMod implements WorldPersistenceHooks.WorldPersistenceHook
     {
         DataGenerator gen = event.getGenerator();
 
-        if (event.includeClient())
-        {
-            gen.addProvider(new ItemModels(gen, event.getExistingFileHelper()));
-            gen.addProvider(new BlockStates(gen, event.getExistingFileHelper()));
-        }
         if (event.includeServer())
         {
             gen.addProvider(new ForgeBlockTagsProvider(gen));
@@ -205,127 +188,6 @@ public class ForgeMod implements WorldPersistenceHooks.WorldPersistenceHook
             gen.addProvider(new ForgeRecipeProvider(gen));
         }
     }
-    
-    public static class ItemModels extends ModelProvider<ItemModelBuilder>
-    {
-        public ItemModels(DataGenerator generator, ExistingFileHelper existingFileHelper)
-        {
-            super(generator, "forge", ITEM_FOLDER, ItemModelBuilder::new, existingFileHelper);
-        }
-        
-        @Override
-        protected void registerModels()
-        {
-            getBuilder("test_generated_model")
-                    .parent(new UncheckedModelFile("item/generated"))
-                    .texture("layer0", new ResourceLocation("block/stone"));
-
-            getBuilder("test_block_model")
-                    .parent(getExistingFile("block/block"))
-                    .texture("all", new ResourceLocation("block/dirt"))
-                    .texture("top", new ResourceLocation("block/stone"))
-                    .element()
-                        .cube("#all")
-                        .face(Direction.UP)
-                            .texture("#top")
-                            .end()
-                        .end();
-        }
-
-        @Override
-        public String getName()
-        {
-            return "Forge Test Item Models";
-        }
-    }
-
-   public static class BlockStates extends BlockstateProvider {
-
-       public BlockStates(DataGenerator gen, ExistingFileHelper exFileHelper) {
-           super(gen, "forge", exFileHelper);
-       }
-
-       @Override
-       protected void registerStatesAndModels() {
-           ModelFile acaciaFenceGate = getBuilder("acacia_fence_gate")
-                   .parent(getExistingFile("block/template_fence_gate"))
-                   .texture("texture", new ResourceLocation("block/acacia_planks"));
-           ModelFile acaciaFenceGateOpen = getBuilder("acacia_fence_gate_open")
-                   .parent(getExistingFile("block/template_fence_gate_open"))
-                   .texture("texture", new ResourceLocation("block/acacia_planks"));
-           ModelFile acaciaFenceGateWall = getBuilder("acacia_fence_gate_wall")
-                   .parent(getExistingFile("block/template_fence_gate_wall"))
-                   .texture("texture", new ResourceLocation("block/acacia_planks"));
-           ModelFile acaciaFenceGateWallOpen = getBuilder("acacia_fence_gate_wall_open")
-                   .parent(getExistingFile("block/template_fence_gate_wall_open"))
-                   .texture("texture", new ResourceLocation("block/acacia_planks"));
-           ModelFile invisbleModel = new UncheckedModelFile(new ResourceLocation("builtin/generated"));
-           VariantBlockstate builder = getVariantBuilder(Blocks.ACACIA_FENCE_GATE);
-           for (Direction dir : FenceGateBlock.HORIZONTAL_FACING.getAllowedValues()) {
-               int angle = (int) dir.getHorizontalAngle();
-               builder
-                       .partialState()
-                            .with(FenceGateBlock.HORIZONTAL_FACING, dir)
-                            .with(FenceGateBlock.IN_WALL, false)
-                            .with(FenceGateBlock.OPEN, false)
-                            .modelForState()
-                                .modelFile(invisbleModel)
-                                .weight(1)
-                            .nextModel()
-                                .modelFile(acaciaFenceGate)
-                                .rotationY(angle)
-                                .uvLock(true)
-                                .weight(100)
-                            .addModel()
-                       .partialState()
-                            .with(FenceGateBlock.HORIZONTAL_FACING, dir)
-                            .with(FenceGateBlock.IN_WALL, false)
-                            .with(FenceGateBlock.OPEN, true)
-                            .modelForState()
-                                .modelFile(acaciaFenceGateOpen)
-                                .rotationY(angle)
-                                .uvLock(true)
-                            .addModel()
-                       .partialState()
-                            .with(FenceGateBlock.HORIZONTAL_FACING, dir)
-                            .with(FenceGateBlock.IN_WALL, true)
-                            .with(FenceGateBlock.OPEN, false)
-                            .modelForState()
-                                .modelFile(acaciaFenceGateWall)
-                                .rotationY(angle)
-                                .uvLock(true)
-                            .addModel()
-                       .partialState()
-                            .with(FenceGateBlock.HORIZONTAL_FACING, dir)
-                            .with(FenceGateBlock.IN_WALL, true)
-                            .with(FenceGateBlock.OPEN, true)
-                            .modelForState()
-                                .modelFile(acaciaFenceGateWallOpen)
-                                .rotationY(angle)
-                                .uvLock(true)
-                            .addModel();
-           }
-
-           ModelFile acaciaFencePost = getBuilder("acacia_fence_post")
-                   .parent(getExistingFile("block/fence_post"))
-                   .texture("texture", new ResourceLocation("block/acacia_planks"));
-           
-           ModelFile acaciaFenceSide = getBuilder("acacia_fence_side")
-                   .parent(getExistingFile("block/fence_side"))
-                   .texture("texture", new ResourceLocation("block/acacia_planks"));
-           
-           getMultipartBuilder(Blocks.ACACIA_FENCE)
-                   .part().modelFile(acaciaFencePost).addModel().build()
-                   .part().modelFile(acaciaFenceSide).uvLock(true).addModel()
-                           .condition(FenceBlock.NORTH, true).build()
-                   .part().modelFile(acaciaFenceSide).rotationY(90).uvLock(true).addModel()
-                           .condition(FenceBlock.EAST, true).build()
-                   .part().modelFile(acaciaFenceSide).rotationY(180).uvLock(true).addModel()
-                           .condition(FenceBlock.SOUTH, true).build()
-                   .part().modelFile(acaciaFenceSide).rotationY(270).uvLock(true).addModel()
-                           .condition(FenceBlock.WEST, true).build();
-       }
-   }
 
     @SubscribeEvent //ModBus, can't use addListener due to nested genetics.
     public void registerRecipeSerialziers(RegistryEvent.Register<IRecipeSerializer<?>> event)
